chore(): partial fixes to hooks
