+++ sdk/package.json
@@ -29,9 +29,9 @@
     "eth-ens-namehash": "2.0.8",
     "ethers": "5.7.2",
     "graphql": "16.7.1",
-    "graphql-request": "5.0.0",
+    "graphql-request": "5.2.0",
     "graphql-tag": "2.12.6",
-    "idb-keyval": "6.2.0",
+    "idb-keyval": "6.2.1",
     "inversify": "6.0.1",
     "ipfs-utils": "9.0.9",
     "key-did-provider-ed25519": "^2.0.1",
