+++ sdk/src/container.ts
@@ -2,7 +2,6 @@ import { Container } from 'inversify';
 import { TYPES } from '@akashaorg/typings/sdk';
 import Logging from './logging';
 import Gql from './gql';
-import GqlNew from './gql/index.new';
 import Settings from './settings';
 import Stash from './stash';
 import DB from './db';
@@ -43,7 +42,6 @@ diContainer.bind<AppSettings>(TYPES.AppSettings).to(AppSettings);
 diContainer.bind<AWF_IC_REGISTRY>(TYPES.ICRegistry).to(AWF_IC_REGISTRY);
 diContainer.bind<AWF_Lit>(TYPES.Lit).to(AWF_Lit);
 diContainer.bind<AWF_Ceramic>(TYPES.Ceramic).to(AWF_Ceramic);
-diContainer.bind(TYPES.GqlNew).to(GqlNew);
 //@Todo: implement init watcher to prevent ambiguous service
 // export const importLazy = async () => {
 //   const r = await import('./gql/index.new');
