+++ sdk/src/gql/index.ts
@@ -1,18 +1,11 @@
 import { inject, injectable } from 'inversify';
 import { TYPES } from '@akashaorg/typings/sdk';
 import Stash, { IQuickLRU } from '../stash';
-import { GraphQLClient } from 'graphql-request';
-import { getSdk, Sdk, SdkFunctionWrapper } from './api';
+import { getSdk, Requester, Sdk } from './api';
 import Logging from '../logging/index';
 import pino from 'pino';
-
-export interface GqlOperation {
-  query: string;
-  variables?: Record<string, unknown>;
-  operationName?: string;
-  context?: Record<string, unknown>;
-  extensions?: Record<string, unknown>;
-}
+import CeramicService from '../common/ceramic';
+import type { DocumentNode } from 'graphql';
 
 /** @internal  */
 @injectable()
@@ -20,21 +13,36 @@ class Gql {
   private _stash: Stash;
 
   readonly _client: Sdk;
+  readonly _ceramic: CeramicService;
   // #_clientWithCache: Sdk;
   private _log: pino.Logger;
   private _gqlStash: IQuickLRU;
 
   // create Apollo link object and initialize the stash
-  public constructor(@inject(TYPES.Stash) stash: Stash, @inject(TYPES.Log) log: Logging) {
+  public constructor(
+    @inject(TYPES.Stash) stash: Stash,
+    @inject(TYPES.Log) log: Logging,
+    @inject(TYPES.Ceramic) ceramic: CeramicService,
+  ) {
     this._stash = stash;
-    const endPoint = new GraphQLClient(process.env.GRAPHQL_URI || '/graphql');
     this._log = log.create('AWF_GQL');
     // create the cache stash for the gql client
     this._gqlStash = this._stash.create({
       maxSize: 1280,
       maxAge: 1000 * 60 * 2,
     }).data;
-    this._client = getSdk(endPoint);
+    this._ceramic = ceramic;
+    const requester = async <R, V>(doc: DocumentNode, vars: V): Promise<R> => {
+      const result = await this._ceramic
+        .getComposeClient()
+        .execute(doc, vars as Record<string, unknown> | undefined);
+      if (!result.errors || !result.errors.length) {
+        return result.data as R;
+      }
+      throw result.errors;
+    };
+
+    this._client = getSdk(requester);
   }
 
   /**
@@ -50,6 +58,7 @@ class Gql {
 
   /**
    * @param withCache - if the client should cache all the requests
+   * remove 'any' from return after all the graphql fragments are updated
    */
   getAPI(withCache = false) {
     return this._client;
