+++ sdk/src/posts/comments.ts
@@ -42,159 +42,159 @@ class AWF_Comments {
     this._auth = auth;
     this._globalChannel = globalChannel;
   }
-
-  /**
-   * Get comment data
-   */
-  @validate(CommentIDSchema)
-  async getComment(commentID: CommentID) {
-    return this._gql.getAPI().GetComment({ commentID: commentID });
-  }
-
-  /**
-   * Get a list of comments for a post
-   */
-  @validate(z.object({ offset: CommentIDSchema, limit: z.number(), postID: EntryIDSchema }))
-  async getComments(opt: { offset?: CommentID; limit: number; postID: EntryID }) {
-    return this._gql
-      .getAPI()
-      .GetComments({ offset: opt.offset, limit: opt.limit, postID: opt.postID });
-  }
-
-  /**
-   * Get a list of replies for a comment
-   */
-  @validate(
-    z.object({
-      offset: CommentIDSchema,
-      limit: z.number(),
-      postID: EntryIDSchema,
-      commentID: CommentIDSchema,
-    }),
-  )
-  async getReplies(opt: {
-    offset?: CommentID;
-    limit: number;
-    postID: EntryID;
-    commentID: CommentID;
-  }) {
-    return this._gql.getAPI().GetReplies({
-      offset: opt.offset,
-      limit: opt.limit,
-      postID: opt.postID,
-      commentID: opt.commentID,
-    });
-  }
-
-  /**
-   * Create a new comment
-   */
-  @validate(
-    z.object({
-      data: z.array(DataProviderInputSchema),
-      comment: z.object({
-        postID: EntryIDSchema,
-        replyTo: CommentIDSchema.optional(),
-        tags: z.array(z.string()).optional(),
-        mentions: z.array(z.string()).optional(),
-      }),
-    }),
-  )
-  async addComment(opt: {
-    data: DataProviderInput[];
-    comment: { postID: EntryID; replyTo?: CommentID; tags?: string[]; mentions?: string[] };
-  }) {
-    const textContent = opt.data.find(e => e.property === 'textContent');
-    if (!textContent) {
-      throwError('Comment does not have content', ['sdk', 'comments', 'addComment']);
-      return;
-    }
-    textContent.value = Buffer.from(textContent.value).toString('base64');
-    const auth = await this._auth.authenticateMutationData(
-      opt.data as unknown as Record<string, unknown>[],
-    );
-    const newComm = await this._gql.getAPI().AddComment(
-      { content: opt.data, comment: opt.comment },
-      {
-        Authorization: `Bearer ${auth.token}`,
-        Signature: auth.signedData.signature.toString(),
-      },
-    );
-    this._globalChannel.next({
-      data: newComm.addComment,
-      event: COMMENTS_EVENTS.CREATE,
-      args: opt,
-    });
-
-    return newComm;
-  }
-
-  /**
-   * Update an existing comment
-   */
-  @validate(
-    z.object({
-      commentID: CommentIDSchema,
-      data: z.array(DataProviderInputSchema),
-      comment: z.object({
-        postID: EntryIDSchema,
-        replyTo: CommentIDSchema.optional(),
-        tags: z.array(z.string()).optional(),
-        mentions: z.array(z.string()).optional(),
-      }),
-    }),
-  )
-  async editComment(opt: {
-    commentID: string;
-    data: DataProviderInput[];
-    comment: { postID: EntryID; replyTo?: CommentID; tags?: string[]; mentions?: string[] };
-  }) {
-    const textContent = opt.data.find(e => e.property === 'textContent');
-    if (!textContent) {
-      throwError('Cannot edit comment with no content', ['sdk', 'comments', 'editComment']);
-      return;
-    }
-    textContent.value = Buffer.from(textContent.value).toString('base64');
-    const auth = await this._auth.authenticateMutationData(
-      opt.data as unknown as Record<string, unknown>[],
-    );
-    const comm = await this._gql.getAPI().EditComment(
-      { content: opt.data, comment: opt.comment, id: opt.commentID },
-      {
-        Authorization: `Bearer ${auth.token}`,
-        Signature: auth.signedData.signature.toString(),
-      },
-    );
-    // @emits COMMENTS_EVENTS.EDIT
-    this._globalChannel.next({
-      data: comm.editComment,
-      event: COMMENTS_EVENTS.EDIT,
-      args: opt,
-    });
-    return comm;
-  }
-
-  /**
-   * Remove a comment's data by ID
-   */
-  @validate(CommentIDSchema)
-  async removeComment(commentID: CommentID) {
-    const auth = await this._auth.authenticateMutationData(commentID);
-    const res = await this._gql.getAPI().RemoveComment(
-      { id: commentID },
-      {
-        Authorization: `Bearer ${auth.token}`,
-        Signature: auth.signedData.signature.toString(),
-      },
-    );
-    // @emits COMMENTS_EVENTS.REMOVE
-    this._globalChannel.next({
-      data: res.removeComment,
-      event: COMMENTS_EVENTS.REMOVE,
-      args: { commentID },
-    });
-    return res;
-  }
+  //
+  // /**
+  //  * Get comment data
+  //  */
+  // @validate(CommentIDSchema)
+  // async getComment(commentID: CommentID) {
+  //   return this._gql.getAPI().GetComment({ commentID: commentID });
+  // }
+  //
+  // /**
+  //  * Get a list of comments for a post
+  //  */
+  // @validate(z.object({ offset: CommentIDSchema, limit: z.number(), postID: EntryIDSchema }))
+  // async getComments(opt: { offset?: CommentID; limit: number; postID: EntryID }) {
+  //   return this._gql
+  //     .getAPI()
+  //     .GetComments({ offset: opt.offset, limit: opt.limit, postID: opt.postID });
+  // }
+  //
+  // /**
+  //  * Get a list of replies for a comment
+  //  */
+  // @validate(
+  //   z.object({
+  //     offset: CommentIDSchema,
+  //     limit: z.number(),
+  //     postID: EntryIDSchema,
+  //     commentID: CommentIDSchema,
+  //   }),
+  // )
+  // async getReplies(opt: {
+  //   offset?: CommentID;
+  //   limit: number;
+  //   postID: EntryID;
+  //   commentID: CommentID;
+  // }) {
+  //   return this._gql.getAPI().GetReplies({
+  //     offset: opt.offset,
+  //     limit: opt.limit,
+  //     postID: opt.postID,
+  //     commentID: opt.commentID,
+  //   });
+  // }
+  //
+  // /**
+  //  * Create a new comment
+  //  */
+  // @validate(
+  //   z.object({
+  //     data: z.array(DataProviderInputSchema),
+  //     comment: z.object({
+  //       postID: EntryIDSchema,
+  //       replyTo: CommentIDSchema.optional(),
+  //       tags: z.array(z.string()).optional(),
+  //       mentions: z.array(z.string()).optional(),
+  //     }),
+  //   }),
+  // )
+  // async addComment(opt: {
+  //   data: DataProviderInput[];
+  //   comment: { postID: EntryID; replyTo?: CommentID; tags?: string[]; mentions?: string[] };
+  // }) {
+  //   const textContent = opt.data.find(e => e.property === 'textContent');
+  //   if (!textContent) {
+  //     throwError('Comment does not have content', ['sdk', 'comments', 'addComment']);
+  //     return;
+  //   }
+  //   textContent.value = Buffer.from(textContent.value).toString('base64');
+  //   const auth = await this._auth.authenticateMutationData(
+  //     opt.data as unknown as Record<string, unknown>[],
+  //   );
+  //   const newComm = await this._gql.getAPI().AddComment(
+  //     { content: opt.data, comment: opt.comment },
+  //     {
+  //       Authorization: `Bearer ${auth.token}`,
+  //       Signature: auth.signedData.signature.toString(),
+  //     },
+  //   );
+  //   this._globalChannel.next({
+  //     data: newComm.addComment,
+  //     event: COMMENTS_EVENTS.CREATE,
+  //     args: opt,
+  //   });
+  //
+  //   return newComm;
+  // }
+  //
+  // /**
+  //  * Update an existing comment
+  //  */
+  // @validate(
+  //   z.object({
+  //     commentID: CommentIDSchema,
+  //     data: z.array(DataProviderInputSchema),
+  //     comment: z.object({
+  //       postID: EntryIDSchema,
+  //       replyTo: CommentIDSchema.optional(),
+  //       tags: z.array(z.string()).optional(),
+  //       mentions: z.array(z.string()).optional(),
+  //     }),
+  //   }),
+  // )
+  // async editComment(opt: {
+  //   commentID: string;
+  //   data: DataProviderInput[];
+  //   comment: { postID: EntryID; replyTo?: CommentID; tags?: string[]; mentions?: string[] };
+  // }) {
+  //   const textContent = opt.data.find(e => e.property === 'textContent');
+  //   if (!textContent) {
+  //     throwError('Cannot edit comment with no content', ['sdk', 'comments', 'editComment']);
+  //     return;
+  //   }
+  //   textContent.value = Buffer.from(textContent.value).toString('base64');
+  //   const auth = await this._auth.authenticateMutationData(
+  //     opt.data as unknown as Record<string, unknown>[],
+  //   );
+  //   const comm = await this._gql.getAPI().EditComment(
+  //     { content: opt.data, comment: opt.comment, id: opt.commentID },
+  //     {
+  //       Authorization: `Bearer ${auth.token}`,
+  //       Signature: auth.signedData.signature.toString(),
+  //     },
+  //   );
+  //   // @emits COMMENTS_EVENTS.EDIT
+  //   this._globalChannel.next({
+  //     data: comm.editComment,
+  //     event: COMMENTS_EVENTS.EDIT,
+  //     args: opt,
+  //   });
+  //   return comm;
+  // }
+  //
+  // /**
+  //  * Remove a comment's data by ID
+  //  */
+  // @validate(CommentIDSchema)
+  // async removeComment(commentID: CommentID) {
+  //   const auth = await this._auth.authenticateMutationData(commentID);
+  //   const res = await this._gql.getAPI().RemoveComment(
+  //     { id: commentID },
+  //     {
+  //       Authorization: `Bearer ${auth.token}`,
+  //       Signature: auth.signedData.signature.toString(),
+  //     },
+  //   );
+  //   // @emits COMMENTS_EVENTS.REMOVE
+  //   this._globalChannel.next({
+  //     data: res.removeComment,
+  //     event: COMMENTS_EVENTS.REMOVE,
+  //     args: { commentID },
+  //   });
+  //   return res;
+  // }
 }
 
 export default AWF_Comments;
