+++ ui/apps/akasha/src/components/item-page/post-page.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import BaseEntryPage from './common/base-page';
 import { useParams } from 'react-router-dom';
 import { RootComponentProps, ModalNavigationOptions, EntityTypes } from '@akashaorg/typings/ui';
-import { usePost, mapEntry } from '@akashaorg/ui-awf-hooks';
 
 type PostPageProps = {
   userId?: string;
@@ -14,17 +13,10 @@ const PostPage: React.FC<PostPageProps & RootComponentProps> = props => {
 
   const { postId } = useParams<{ postId: string }>();
   // @TODO replace with new hooks
-  const postReq = usePost({
-    postId,
-    loggedUser: userId,
-  });
-
+  const postReq = undefined;
   const entryData = React.useMemo(() => {
-    if (postReq.data) {
-      return mapEntry(postReq.data);
-    }
     return undefined;
-  }, [postReq.data]);
+  }, []);
 
   return (
     <BaseEntryPage
