+++ ui/apps/akasha/src/extensions/inline-editor/reply-editor/index.tsx
@@ -1,13 +1,6 @@
 import * as React from 'react';
 
 import { RootExtensionProps, IPublishData, AnalyticsCategories } from '@akashaorg/typings/ui';
-import {
-  useEditComment,
-  mapEntry,
-  useAnalytics,
-  useComment,
-  useCreateComment,
-} from '@akashaorg/ui-awf-hooks';
 import { useTranslation } from 'react-i18next';
 import { Base } from '../base';
 import EntryCardLoading from '@akashaorg/design-system-components/lib/components/Entry/EntryCardLoading';
@@ -20,19 +13,16 @@ type Props = {
 
 export function ReplyEditor({ commentId, singleSpa, action }: Props) {
   const { t } = useTranslation('app-akasha-integration');
-  const [analyticsActions] = useAnalytics();
+  const [analyticsActions] = [undefined];
 
   // @TODO replace with new hooks
-  const comment = useComment(commentId, true);
-  const editComment = useEditComment(commentId, true);
-  const publishComment = useCreateComment();
+  const comment = undefined;
+  const editComment = undefined;
+  const publishComment = undefined;
 
   const entryData = React.useMemo(() => {
-    if (comment.status === 'success') {
-      return mapEntry(comment.data);
-    }
     return undefined;
-  }, [comment.data, comment.status]);
+  }, []);
 
   const [editorState, setEditorState] = React.useState(
     action === 'edit' ? entryData?.slateContent : null,
