+++ ui/design-system-components/src/components/ProfileEngagements/types.ts
@@ -1,12 +1,12 @@
-import { Follow, ProfileImageVersions } from '@akashaorg/typings/sdk/graphql-types-new';
+import { AkashaFollow, AkashaProfileImageVersions } from '@akashaorg/typings/sdk/graphql-types-new';
 import { Profile } from '@akashaorg/typings/ui';
 import { ReactElement } from 'react';
 
 export const LOADING_LIST_SIZE = 10;
 
 export type Engagement = {
-  id: Follow['id'];
-  isFollowing: Follow['isFollowing'];
+  id: AkashaFollow['id'];
+  isFollowing: AkashaFollow['isFollowing'];
   profile?: Profile;
 }[];
 
@@ -14,7 +14,7 @@ export type EngagementProps = {
   profileAnchorLink: string;
   loadMore: boolean;
   onLoadMore: () => void;
-  getMediaUrl: (image?: ProfileImageVersions) => ProfileImageVersions;
+  getMediaUrl: (image?: AkashaProfileImageVersions) => AkashaProfileImageVersions;
   renderFollowElement: (
     profileStreamId: string,
     followStreamId: string,
