+++ ui/design-system-components/src/components/types/common.types.ts
@@ -1,7 +1,7 @@
-import { ProfileLinkSource } from '@akashaorg/typings/sdk/graphql-types-new';
+import { AkashaProfileLinkSource } from '@akashaorg/typings/sdk/graphql-types-new';
 import { Profile } from '@akashaorg/typings/ui';
 
-export type Link = ProfileLinkSource;
+export type Link = AkashaProfileLinkSource;
 
 export type Developer = {
   profileId: Profile['did']['id'];
