+++ ui/design-system-core/src/utils/dummy-data.ts
@@ -1,4 +1,4 @@
-import { FollowConnection, Profile } from '@akashaorg/typings/sdk/graphql-types-new';
+import { AkashaFollowConnection, AkashaProfile } from '@akashaorg/typings/sdk/graphql-types-new';
 
 const userAvatar = (width = 360, height = 360) => {
   return {
@@ -20,10 +20,10 @@ const createUser = n => {
     },
     avatar: userAvatar(),
     createdAt: '2021-03-01T00:00:00.000Z',
-    followers: {} as FollowConnection,
+    followers: {} as AkashaFollowConnection,
   };
 };
-const userData: Profile[] = Array.from({ length: 4 }, (v, k) => createUser(k));
+const userData: AkashaProfile[] = Array.from({ length: 4 }, (v, k) => createUser(k));
 
 const sidebarItems = [
   {
@@ -55,14 +55,14 @@ const sidebarItems = [
   },
 ];
 
-const profileData: Profile = {
+const profileData: AkashaProfile = {
   did: { id: 'did:pkh:eip155:1:0x003410490050000320006570034567114572000', isViewer: true },
   avatar: userAvatar(),
   background: userAvatar(540, 320),
   name: 'Gilbert The Bearded',
   description:
     'Product design @companyname. Main interests: User experience, Design processes, Project Management. Author of This could be a book name, and Another Book. Love people, plants, words, and food.',
-  followers: {} as FollowConnection,
+  followers: {} as AkashaFollowConnection,
   createdAt: '2023-01-01T00:00:00.000Z',
   id: '31231',
 };
