+++ ui/webpack.config.js
@@ -31,7 +31,7 @@ const exp = {
     ],
   },
   resolve: {
-    extensions: ['.ts', '.tsx', '.js', '.jsx'],
+    extensions: ['.ts', '.tsx', '.js', '.jsx', '.mjs'],
   },
   output: {
     libraryTarget: 'system',
